Conflicts:
	cities_light/migrations/0001_initial.py
	cities_light/migrations/__init__.py
	cities_light/south_migrations/0002_city_latitude_and_longitude_added.py
	cities_light/south_migrations/0003_auto__add_field_city_search_names.py
	cities_light/south_migrations/0004_added_region.py
	cities_light/south_migrations/0005_set_region.py
	cities_light/south_migrations/0006_add_city_alternate_names.py
	cities_light/south_migrations/0007_region_geoname_id_to_geoname_code.py
	cities_light/south_migrations/0008_add_region_geoname_id.py
	cities_light/south_migrations/0009_alternate_names_for_all.py
	cities_light/south_migrations/0010_set_geoname_ids.py
	cities_light/south_migrations/0011_add_city_and_region_display_name.py
	cities_light/south_migrations/0012_set_display_name.py
	cities_light/south_migrations/0013_geoname_id_unique_index.py
	cities_light/south_migrations/0014_auto__chg_field_city_search_names.py
	cities_light/south_migrations/0015_auto__chg_field_country_slug__add_field_city_population__add_field_cit.py

+++ test.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 set -x
+set -e
 
 function do_db() {
     python test_project/manage.py syncdb --traceback --noinput --settings=test_project.$1
@@ -7,17 +8,25 @@ function do_db() {
     python test_project/manage.py cities_light --force-import-all --traceback --settings=test_project.$1
 }
 
-pip install south psycopg2
+pip install south 
 
-if [[ ${TRAVIS_PYTHON_VERSION%%.*} -eq "2" ]]; then
+if [[ $DB = 'mysql' ]]; then
+    if [[ $TRAVIS_PYTHON_VERSION == 2* ]]; then
+        pip install mysql-python
+    else
+        pip install -e git+https://github.com/clelland/MySQL-for-Python-3.git#egg=mysql-python
+    fi
+
     # test on mysql
-    pip install mysql-python 
     do_db settings_mysql
 fi
 
-# test on postgres
-do_db settings_postgres
+if [[ $DB = 'postgresql' ]]; then
+    pip install psycopg2
+    do_db settings_postgres
+fi 
 
-# test on sqlite
-rm -rf test_project/db.sqlite
-do_db settings
+if [[ $DB = 'sqlite' ]]; then
+    rm -rf test_project/db.sqlite
+    do_db settings
+fi
